getcwd, libc, функция process_open_file_relative

+++ sdk/libc/src/process.c
@@ -30,7 +30,7 @@ void sleep(int time)
 void* mmap(void* addr, size_t length, int protection, int flags)
 {
     void* result = syscall_process_map_memory(addr, length, protection, flags);
-    int rc = result;
+    int64_t rc = (int64_t) result;
     if (rc < 0) { 
         errno = -rc; 
         return (void*)-1;
